Reorganizing secondary tools.
Only the "main" commands of the bcbgo package should be in the 'cmd'
directory. All other tools (mostly stored as "examples" before this)
should be in the 'tools' directory. Tools are not required to use bcbgo,
but are typically useful for intermediate tasks and are necessary for
running experiments.

Also, I've fixed the output of HHM files (including the MSA part) so
that they can be read by HHfrag's parser. Ticky tack!